% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutomaticDemandIdentifier.R
\name{adi}
\alias{adi}
\title{Automatic Demand Identifier}
\usage{
adi(y, ic = c("AIC", "AICc", "BIC", "BICc"))
}
\arguments{
\item{y}{The vector of the data.}

\item{ic}{Information criterion to use.}
}
\value{
Class "adi" is returned, which contains:
\itemize{
\item models - All fitted models;
\item ICs - Values of information criteria;
\item type - The type of the identified demand.
}
}
\description{
The function applies several models on the provided time series and identifies what
type of demand it is based on an information criterion.
}
\details{
The function creates explanatory variables based on LOWESS of the original data, then
applies Normal, Normal + Bernoulli, Negative Binomial and Negative Binomial + Bernoulli
models and selects the one that has the lowest IC. Based on that, it decides what type
of demand the data corresponds to: regular non-count, intermittent non-count,
regular count, or intermittent count.
}
\examples{
# Data from Poisson distribution
y <- rpois(120, 0.7)
adi(y)

}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}
}
\keyword{models}
\keyword{nonlinear}
\keyword{regression}
\keyword{ts}
