% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combiner.R
\name{combiner}
\alias{combiner}
\title{Combine regressions based on information criteria}
\usage{
combiner(data, ic = c("AICc", "AIC", "BIC"), silent = TRUE)
}
\arguments{
\item{data}{Data frame containing dependant variable in the first column and
the others in the rest.}

\item{ic}{Information criterion to use.}

\item{silent}{If \code{silent=FALSE}, then nothing is silent, everything is
printed out. \code{silent=TRUE} means that nothing is produced.}
}
\value{
Function returns \code{model} - the final model of the class
"lm.combined".
}
\description{
Function combines parameters of linear regressions of the first variable
on all the other provided data.
}
\details{
The algorithm uses lm() to fit different models and then combines the models
based on the selected IC.
}
\examples{

xreg <- cbind(rnorm(100,10,3),rnorm(100,50,5))
xreg <- cbind(100+0.5*xreg[,1]-0.75*xreg[,2]+rnorm(100,0,3),xreg,rnorm(100,300,10))
colnames(xreg) <- c("y","x1","x2","Noise")

combiner(xreg)

}
\references{
\itemize{
\item Burnham Kenneth P. and Anderson David R. (2002). Model Selection
and Multimodel Inference. A Practical Information-Theoretic Approach.
Springer-Verlag New York. \url{http://dx.doi.org/10.1007/b97636}.
}
}
\seealso{
\code{\link[stats]{step}, \link[greybox]{xregExpander},
\link[greybox]{stepwise}}
}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}
}
\keyword{models}
\keyword{nonlinear}
\keyword{regression}
\keyword{ts}
