% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr.R
\name{folder}
\alias{folder}
\title{Create a dataset from the provided data and models}
\usage{
folder(...)
}
\arguments{
\item{...}{Objects that need to be merged in one dataset.}
}
\value{
Function returns \code{data} - the data.frame of the variables,
\code{objects} - the list of all provided objects and \code{dataType} - vector
with types of objects provided to folder.
}
\description{
Function creates a data.frame from the objects provided to it. The objects
include vector, matrix, data.frame and any model that supports the standard
methods (such as fitted, vcov, predict, forecast). In the latter case,
the function will extract fitted values and use them instead of the actual one.
}
\details{
The function is needed as a first step in the estimation of a regression model
with stochastic regressors.
}
\examples{

### Simple example
xreg <- cbind(rnorm(100,10,3),rnorm(100,50,5))
xreg <- cbind(100+0.5*xreg[,1]-0.75*xreg[,2]+rnorm(100,0,3),xreg,rnorm(100,300,10))
colnames(xreg) <- c("y","x1","x2","Noise")
modelForY <- stepwise(xreg)

foldedData <- folder(y=xreg[,4], xreg[,-c(1,4)], z=modelForY)
vcov(foldedData)

}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}
}
\keyword{models}
\keyword{nonlinear}
\keyword{regression}
\keyword{ts}
