% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc.R
\name{icc}
\alias{icc}
\title{Intraclass correlation}
\usage{
icc(x, y)
}
\arguments{
\item{x}{Either factor or a matrix with dummies}

\item{y}{The numerical variable.}
}
\value{
The following list of values is returned:
\itemize{
\item{value}{The value of the coefficient;}
\item{statistic}{The value of F-statistics associated with the parameter;}
\item{p.value}{The p-value of F-statistics associated with the parameter;}
\item{df.residual}{The number of degrees of freedom for the residuals;}
\item{df}{The number of degrees of freedom for the data.}
}
}
\description{
Function calculates Intraclass correlation for two variables
}
\details{
This is based on the linear regression model with the set of dummies providede in x.
The returned value is just a coefficient of multiple correlation from such a model,
the F-statistics of the model (thus testing the null hypothesis that all the
parameters are equal to zero), the associated p-value and the degrees of freedom.
}
\examples{

icc(mtcars$am, mtcars$mpg)

}
\seealso{
\code{\link[base]{table}, \link[greybox]{tableplot}, \link[greybox]{spread},
\link[greybox]{cramer}}
}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}
}
\keyword{models}
\keyword{nonlinear}
\keyword{regression}
\keyword{ts}
