% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemenyi.R
\name{nemenyi}
\alias{nemenyi}
\title{Nemenyi test}
\usage{
nemenyi(data, level = 0.95, sort = c(TRUE, FALSE), type = c("vline",
  "none", "mcb", "vmcb", "line"), select = NULL, labels = NULL, ...)
}
\arguments{
\item{data}{Matrix or data frame with observations in rows and variables in
columns.}

\item{level}{The width of the confidence interval. Default is 0.95.}

\item{sort}{If \code{TRUE} function sorts the final values of mean ranks.
If plots are requested via \code{type} parameter, then this is forced to
\code{TRUE}.}

\item{type}{Type of the plot. This can be: 1. "none" - No plot;
2. "mcb" - MCB style plot; 3. "vmcb" - Vertical MCB style plot;
4. "line" - Line visualisation (ISF style), where numbers next to method
names are the mean ranks; 5. "vline" - Vertical line visualisation.}

\item{select}{Highlight the selected model (column). Number 1 to ncol. If
\code{NULL}, then no highlighting is done.}

\item{labels}{Optional labels for models. If \code{NULL} names of columns of
\code{data} will be used. If number of labels != ncol, then it is assumed to
be \code{NULL}.}

\item{...}{Other parameters passed to plot function.}
}
\value{
Function returns the following variables:
\itemize{
\item{means}{Mean rank of each method.}
\item{intervals}{Nemenyi intervals for each method.}
\item{p.value}{Friedman test p-value.}
\item{hypothesis}{Friedman hypothesis result.}
\item{CD}{Critical distance for nemenyi.}
\item{level}{Singificance level.}
\item{nMethods}{Number of methods.}
\item{obs}{Number of measurements.}
}
}
\description{
Nemenyi is a non-parametric test for the comparison of medians of several
distributions
}
\details{
Hibon et al.(2012) demonstrated that Nemenyi test is equivalent to MCB
(Multiple Comparison with the Best). The test ranks values of the variables
and constructs confidence intervals. Based on these intervals the grouping of
distributions is done, and a graph is produced so that it becomes apparent
if there is a statistically significant difference between them.
}
\examples{

N <- 50
M <- 4
ourData <- matrix(rnorm(N*M,mean=0,sd=1), N, M)
ourData[,2] <- ourData[,2]+1
ourData[,3] <- ourData[,3]+0.7
ourData[,4] <- ourData[,4]+0.5
colnames(ourData) <- c("Method A","Method B","Method C - long name","Method D");
nemenyi(ourData, level=0.95, type="vline")

par(mar=c(2,0,2,0),cex=1.5)
nemenyi(ourData,level=0.95,type="vline",main="",xlim=range(0,1.5))

}
\references{
\itemize{
\item Koning, A. J., Franses, P. H., Hibon, M., & Stekler, H. O.
(2005). The M3 competition: Statistical tests of the results.
International Journal of Forecasting, 21(3), 397-409.
\url{https://doi.org/10.1016/j.ijforecast.2004.10.003}
\item Demsar, J. (2006). Statistical Comparisons of Classifiers over
Multiple Data Sets. Journal of Machine Learning Research, 7, 1-30.
\url{http://www.jmlr.org/papers/volume7/demsar06a/demsar06a.pdf}
\item Hibon M., Crone S. F., Kourentzes N., 2012. Statistical
significance of forecasting methods: An empirical evaluation of the
robustness and interpretability of the MCB, ANOM and Nemenyi tests,
The 32nd Annual International Symposium on Forecasting, Boston.
}
}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}

Nikolaos Kourentzes
}
\keyword{htest}
