% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.greybox}
\alias{plot.greybox}
\title{Plots for the fit and residuals}
\usage{
\method{plot}{greybox}(x, which = c(1, 2, 3, 4), level = 0.95, ...)
}
\arguments{
\item{x}{Time series model for which forecasts are required.}

\item{which}{Forecast horizon}

\item{level}{Confidence level. Defines width of prediction interval.}

\item{...}{The parameters passed to the plot functions.}
}
\value{
The function produces 4 plots and, if \code{any(which==3)} also prints the number
of residuals otside the bounds.
}
\description{
The function produces fitted values and plots for the residuals of the greybox functions
}
\details{
The list of produced plots includes:
\enumerate{
\item Fitted over time;
\item Standardised residuals vs Fitted;
\item Absolute residuals vs Fitted;
\item Q-Q plot with the specified distribution.
}
Which of the plots to produce, is specified via the \code{which} parameter. The first
two plots also use the parameters \code{level}, which specifies the confidence level for
the intervals.
}
\examples{

xreg <- cbind(rlaplace(100,10,3),rnorm(100,50,5))
xreg <- cbind(100+0.5*xreg[,1]-0.75*xreg[,2]+rlaplace(100,0,3),xreg,rnorm(100,300,10))
colnames(xreg) <- c("y","x1","x2","Noise")

ourModel <- alm(y~x1+x2, xreg, distribution="dlaplace")

plot(ourModel)

}
\seealso{
\link[stats]{plot.lm}
}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}
}
\keyword{ts}
\keyword{univar}
