% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{polyprod}
\alias{polyprod}
\title{This function calculates parameters for the polynomials}
\usage{
polyprod(x, y)
}
\arguments{
\item{x}{The vector of parameters of the first polynomial.}

\item{y}{The vector of parameters of the second polynomial.}
}
\value{
The function returns a matrix with one column with the parameters for
the polynomial, starting from the 0-order.
}
\description{
The function accepts two vectors with the parameters for the polynomials and returns
the vector of parameters after their multiplication. This can be especially useful,
when working with ARIMA models.
}
\examples{

polyprod(c(1,-2,-1),c(1,0.5,0.3))

}
\seealso{
\link[stats]{convolve}
}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}
}
