% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread.R
\name{spread}
\alias{spread}
\title{Construct scatterplot / boxplots for the data}
\usage{
spread(data, histograms = FALSE, ...)
}
\arguments{
\item{data}{Either matrix or data.frame with the data.}

\item{histograms}{If \code{TRUE}, then the histrograms and barplots are produced on
the diagonal of the matrix.}

\item{...}{Other parameters passed to the plot function.}
}
\value{
Function does not return anything. It just plots things.
}
\description{
Function constructs the plots depending on the types of variables in the provided
matrix / data.frame.
}
\details{
If the variables are in metric scale, then the classical scatterplot is constructed.
If they one of them is either integer (up to 10 options) or categorical, then
boxplot is constructed. Finally, for the two categorical variables the table plot is
returned. All of that is packed in a matrix.
}
\examples{

### Simple example
spread(mtcars)

}
\references{
\itemize{
\item Burnham Kenneth P. and Anderson David R. (2002). Model Selection
and Multimodel Inference. A Practical Information-Theoretic Approach.
Springer-Verlag New York. DOI: [10.1007/b97636](http://dx.doi.org/10.1007/b97636).
}
}
\seealso{
\code{\link[graphics]{plot}, \link[base]{table}, \link[greybox]{tableplot}}
}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}
}
\keyword{models}
\keyword{nonlinear}
\keyword{regression}
\keyword{ts}
