% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr.R
\name{srm}
\alias{srm}
\title{Stochastic Regressors Model}
\usage{
srm(formula, folder, subset = NULL, na.action = NULL, ...)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to
that class): a symbolic description of the model to be fitted. Can also include
\code{trend}, which would add the global trend.}

\item{folder}{the \link[greybox]{folder} object.}

\item{subset}{an optional vector specifying a subset of observations to be
used in the fitting process.}

\item{na.action}{a function which indicates what should happen when the
data contain NAs. The default is set by the na.action setting of
\link[base]{options}, and is \link[stats]{na.fail} if that is unset. The
factory-fresh default is \link[stats]{na.omit}. Another possible value
is NULL, no action. Value \link[stats]{na.exclude} can be useful.}

\item{...}{All the other parameters passed to \link[greybox]{alm}.}
}
\value{
Function returns exactly the same set of variables as
\link[greybox]{alm} but with addition of \code{objects} list, containing
all the objects used in the original \link[greybox]{folder} call.
}
\description{
This function is a wrapper for \link[greybox]{alm}, which extracts the design matrix from
the \link[greybox]{folder} object and calls for \code{alm()} to estimate the model on it.
}
\details{
The function accepts all the parameters of \link[greybox]{alm}, so have a look at the
documentation of that function.
}
\examples{
xreg <- cbind(rnorm(100,10+0.5*c(1:100),3),rnorm(100,50,5))
xreg <- cbind(100+0.5*xreg[,1]-0.75*xreg[,2]+rnorm(100,0,3),xreg,rnorm(100,300,10))
colnames(xreg) <- c("y","x1","x2","Noise")

# Estimate the model for the variable x1
modelForX1 <- alm(x1~trend, xreg)

foldedData <- folder(xreg, z=modelForX1)

srModel <- srm(y~x2+z, foldedData)
summary(srModel)

}
\seealso{
\code{\link[greybox]{stepwise}, \link[greybox]{lmCombine},
\link[greybox]{alm}}.
}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}
}
\keyword{models}
\keyword{nonlinear}
\keyword{regression}
\keyword{ts}
