% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableplot.R
\name{tableplot}
\alias{tableplot}
\title{Construct a plot for categorical variable}
\usage{
tableplot(x, y = NULL, labels = TRUE, ...)
}
\arguments{
\item{x}{First categorical variable.}

\item{y}{Second categorical variable.}

\item{labels}{Whether to print table labels inside the plot or not.}

\item{...}{Other parameters passed to the plot function.}
}
\value{
Function does not return anything. It just plots things.
}
\description{
Function constructs a plot for two categorical variables based on table function
}
\details{
The returned plot corresponds to the values of \code{table()} function with
density corresponding to the frequency of appearance. If the value appears more
often than the other (e.g. 0.5 vs 0.15), then it will be darker. The frequency of 0
corresponds to the white colour, the frequency of 1 corresponds to the black.
}
\examples{

tableplot(mtcars$am, mtcars$gear)

}
\seealso{
\code{\link[graphics]{plot}, \link[base]{table}, \link[greybox]{spread}}
}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}
}
\keyword{graph}
\keyword{plots}
