% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefbootstrap.R
\name{timeboot}
\alias{timeboot}
\title{Time series bootstrap}
\usage{
timeboot(y, nsim = 100, scale = NULL, type = c("auto", "multiplicative",
  "additive"), f = 0.002)
}
\arguments{
\item{y}{The original time series}

\item{nsim}{Number of iterations (simulations) to run.}

\item{scale}{Parameter that defines how to scale the variability around the data.}

\item{type}{Type of bootstrap to use. \code{"additive"} means that the randomness is
added, while \code{"multiplicative"} implies the multiplication. By default the function
will try using the latter, unless the data has non-positive values.}

\item{f}{The smoother span parameter for LOWESS (see \link[stats]{lowess}). Small value
should be used to capture the trend in the sorted data correctly.}
}
\value{
The function returns the matrix with the new series in columns and observations
in rows.
}
\description{
The function implements a bootstrap inspired by the Maximum Entropy Bootstrap
}
\details{
The function implements the following algorithm:

1. Sort the original data, recording the original order of elements;
2. Smooth the sorted data using LOWESS to get intermediate points;
3. Take first differences of the sorted series and sort them;
4. Create contingency table based on the differences and take the cumulative sum
of it. This way we end up with an empirical CDF of differences;
5. Generate random numbers from the uniform distribution between 0 and 1;
6. Get the differences that correspond to the random number (randomly extract
empirical quantiles);
7. Add the random differences to the smooth series from (2) to get new series;
8. Reorder (7) based on the initial order of series.

If the multiplicative bootstrap is used then logarithms of the sorted series and LOWESS
are used and at the very end, the exponent of the resulting data is taken. This way the
discrepancies in the data have similar scale no matter what the level of the original
series is. In case of the additive bootstrap, the trended series will be noisier when
the level of series is low.
}
\examples{
plot(AirPassengers, type="l")
timeboot(AirPassengers) |> lines(col="blue")

}
\references{
Vinod HD, López-de-Lacalle J (2009). "Maximum Entropy Bootstrap for Time Series:
The meboot R Package." Journal of Statistical Software, 29(5), 1–19. \doi{doi:10.18637/jss.v029.i05}.
}
\seealso{
\code{\link[stats]{lowess}}
}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}
}
\keyword{models}
\keyword{nonlinear}
\keyword{regression}
\keyword{ts}
