% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alm.R
\name{alm}
\alias{alm}
\title{Advanced Linear Model}
\usage{
alm(formula, data, subset = NULL, na.rm = TRUE,
  distribution = c("dlaplace", "ds", "dfnorm"))
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to
that class): a symbolic description of the model to be fitted.}

\item{data}{a data frame or a matrix, containing the variables in the model.}

\item{subset}{an optional vector specifying a subset of observations to be
used in the fitting process.}

\item{na.rm}{if \code{TRUE}, then observations with missing values are
removed. Otherwise they are interpolated.}
}
\value{
Function returns \code{model} - the final model of the class
"alm".
}
\description{
Function estimates model based on the selected distribution
}
\details{
This is a function, similar to \link[stats]{lm}, but for the cases of several
non-normal distributions. These include:

\enumerate{
\item Laplace distribution,
\item S-distribution,
\item Folded-normal distribution.
}

Some other distributions can be used as well, given that they are formulated
in terms of mean and some other parameters (NOT YET IMPLEMENTED).

The estimation is done using likelihood of respective distributions.
}
\examples{


}
\seealso{
\code{\link[greybox]{stepwise}, \link[greybox]{lmCombine}}
}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}
}
\keyword{models}
\keyword{nonlinear}
\keyword{regression}
\keyword{ts}
