% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determination.R
\name{determination}
\alias{determination}
\title{Determination coefficients}
\usage{
determination(xreg, ...)
}
\arguments{
\item{xreg}{Data frame or a matrix, containing the exogenous variables.}

\item{...}{Other values passed to cor function.}
}
\value{
Function returns the vector of determination coefficients.
}
\description{
Function produces determination coefficient for the provided data
}
\details{
The function calculates determination coefficients (aka R^2)
between all the provided variables. The higher the coefficient is,
the higher the potential multicollinearity effect in the model with
the variables will be. Coefficients of determination are connected
directly to Variance Inflation Factor (VIF): VIF = 1 / (1 -
determination). Arguably it is easier to interpret, because it is
restricted with (0, 1) bounds. The multicollinearity can be
considered as serious, when determination > 0.9 (which corresponds
to VIF > 10).
}
\examples{

### Simple example
xreg <- cbind(rnorm(100,10,3),rnorm(100,50,5))
xreg <- cbind(100+0.5*xreg[,1]-0.75*xreg[,2]+rnorm(100,0,3),xreg,rnorm(100,300,10))
colnames(xreg) <- c("x1","x2","x3","Noise")
determination(xreg)

}
\seealso{
\code{\link[stats]{cor}}
}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}
}
\keyword{models}
