% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplace.R
\name{dlaplace}
\alias{dlaplace}
\alias{plaplace}
\alias{qlaplace}
\alias{rlaplace}
\title{Laplace Distribution}
\usage{
dlaplace(q, mu = 0, b = 1, log = FALSE)

plaplace(q, mu = 0, b = 1)

qlaplace(p, mu = 0, b = 1)

rlaplace(n = 1, mu = 0, b = 1)
}
\arguments{
\item{q}{vector of quantiles.}

\item{mu}{vector of location parameters (means).}

\item{b}{vector of mean absolute errors.}

\item{log}{if \code{TRUE}, then probabilities are returned in
logarithms.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. Should be a single number.}
}
\value{
Depending on the function, various things are returned
(usually either vector or scalar):
\itemize{
\item \code{ds} returns the density function value for the
provided parameters.
\item \code{ps} returns the value of the cumulative function
for the provided parameters.
\item \code{qs} returns quantiles of the distribution. Depending
on what was provided in \code{p}, \code{mu} and \code{b}, this
can be either a vector or a matrix, or an array.
\item \code{rs} returns a vector of random variables
generated from the Laplace distribution. Depending on what was
provided in \code{mu} and \code{b}, this can be either a vector
or a matrix or an array.
}
}
\description{
Density, cumulative distribution, quantile functions and random
generation for the Laplace distribution with the location parameter mu
and Mean Absolute Error (or Mean Absolute Deviation) equal to b.
}
\details{
When mu=0 and b=1, the Laplace distribution becomes standardized with.
The distribution has the following density function:

f(x) = 1/(2b) exp(-abs(x-mu) / b)
}
\examples{
x <- dlaplace(c(-1000:1000)/10, 0, 1)
plot(x, type="l")

x <- plaplace(c(-1000:1000)/10, 0, 1)
plot(x, type="l")

qlaplace(c(0.025,0.975), 0, c(1,2))

x <- rlaplace(1000, 0, 1)
hist(x)

}
\references{
\itemize{
\item Wikipedia page on Laplace distribution:
\url{https://en.wikipedia.org/wiki/Laplace_distribution}.
}
}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}
}
\keyword{distribution}
