% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{getResponse}
\alias{reexports}
\alias{forecast}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{forecast}{\code{\link[forecast]{getResponse}}, \code{\link[forecast]{forecast}}}
}}

