% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeboot.R
\name{timeboot}
\alias{timeboot}
\title{Time series bootstrap}
\usage{
timeboot(y, nsim = 100, intermittent = c("yes", "no"),
  type = c("additive", "multiplicative"), kind = c("nonparametric",
  "parametric"), lag = frequency(y), scale = NULL)
}
\arguments{
\item{y}{The original time series}

\item{nsim}{Number of iterations (simulations) to run.}

\item{intermittent}{Whether to treat the demand as intermittent or not.}

\item{type}{Type of bootstrap to use. \code{"additive"} means that the randomness is
added, while \code{"multiplicative"} implies the multiplication.}

\item{kind}{A kind of the bootstrap to do: nonparametric or parametric. The latter
relies on the normal distribution, while the former uses the empirical distribution of
differences of the data.}

\item{lag}{The lag to use in the calculation of differences. Should be 1 for non-seasonal
data.}

\item{scale}{Scale (sd) to use in the normal distribution. Estimated as mean absolute
differences of the data if omitted.}
}
\value{
The function returns:
\itemize{
\item \code{call} - the call that was used originally;
\item \code{data} - the original data used in the function;
\item \code{boot} - the matrix with the new series in columns and observations in rows.
\item \code{type} - type of the bootstrap used.}
}
\description{
The function implements a bootstrap inspired by the Maximum Entropy Bootstrap
}
\details{
The function implements the following algorithm:

1. Sort the data in the ascending order, recording the original order of elements;
2. Take first differences of the sorted series and sort them;
3. Create contingency table based on the differences and take the cumulative sum
of it. This way we end up with an empirical CDF of differences;
4. Generate random numbers from the uniform distribution between 0 and 1;
5. Get the differences that correspond to the random numbers (randomly extract
empirical quantiles). This way we take the empirical density into account when
selecting the differences;
6. Add the random differences to the sorted series from (1) to get a new time series;
7. Sort the new time series in the ascending order;
8. Reorder (7) based on the initial order of series.

If the multiplicative bootstrap is used then logarithms of the sorted series
are used and at the very end, the exponent of the resulting data is taken. This way the
discrepancies in the data have similar scale no matter what the level of the original
series is. In case of the additive bootstrap, the trended series will be noisier when
the level of series is low.
}
\examples{
timeboot(AirPassengers) |> plot()

}
\references{
Vinod HD, López-de-Lacalle J (2009). "Maximum Entropy Bootstrap for Time Series:
The meboot R Package." Journal of Statistical Software, 29(5), 1–19. \doi{doi:10.18637/jss.v029.i05}.
}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}
}
\keyword{models}
\keyword{nonlinear}
\keyword{regression}
\keyword{ts}
